ci(gcp): add GCP test workflows and convert modules/gcp tests to credential-free
- Add .github/workflows/gcp-tests.yml: runs the credential-free
  modules/gcp suite on every push, no secrets required.
- Add .github/workflows/gcp-integration-tests.yml: runs
  '-tags gcp ./test/gcp/...' via OIDC/Workload Identity Federation,
  gated on GCP_WORKLOAD_IDENTITY_PROVIDER_ID, GCP_SA_EMAIL,
  GCP_PROJECT_ID, GCP_ALLOWED_AUDIENCES.
- Replace each //go:build gcp file in modules/gcp with its
  credential-free *WithClient counterpart so the suite runs without
  any GCP auth.
- Rewrite static_token_test.go to verify GOOGLE_OAUTH_ACCESS_TOKEN
  short-circuits ADC, without hitting real GCP.
